Fix filtering by multi-character primary keys
Filtering by multi-character primary keys is broken in Django 5.0+,
for example with a query string like ?author=11.

This is a bug in the wagtail-modeladmin package that
wagtail-treemodeladmin relies upon. There is an open PR to fix [0] but
until that time we need a workaround.

This is related to this [1] Django issue.

[0] https://github.com/wagtail-nest/wagtail-modeladmin/pull/55
[1] https://code.djangoproject.com/ticket/35316